[refactor/#294-slack] Feign Client 적용한 v1 분리
