--- slack-service/src/main/java/com/business/slackservice/application/dto/response/slackEventType/ResSlackEventTypePostDTOApiV1.java
@@ -1,29 +1,0 @@
-package com.business.slackservice.application.dto.response.slackEventType;
-
-import com.business.slackservice.domain.slackEventType.entity.SlackEventTypeEntity;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-
-@Builder
-public class ResSlackEventTypePostDTOApiV1 {
-    @JsonProperty
-    private SlackEventType slackEventType;
-
-    public static ResSlackEventTypePostDTOApiV1 of(SlackEventTypeEntity slackEventTypeEntity) {
-        return ResSlackEventTypePostDTOApiV1.builder()
-            .slackEventType(SlackEventType.from(slackEventTypeEntity))
-            .build();
-    }
-
-    @Builder
-    public static class SlackEventType {
-        @JsonProperty
-        private String name;
-
-        private static SlackEventType from(SlackEventTypeEntity slackEventTypeEntity) {
-            return SlackEventType.builder()
-                .name(slackEventTypeEntity.getName())
-                .build();
-        }
-    }
-}
